bikeshed/requirements.txt
django
psycopg2
django-crispy-forms
django_resized
Pillow
